requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
braceexpand==0.1.7
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
easy-pil==0.1.9
frozenlist==1.3.3
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
humanfriendly==10.0
idna==3.4
import-expression==1.1.4
jishaku==2.3.0
line-profiler==3.5.1
multidict==6.0.2
nextcord==2.2.0
num2words==0.5.12
Pillow==9.0.1
pip==21.2.3
protobuf==4.21.9
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
setuptools==57.4.0
six==1.16.0
timeago==1.0.16
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wavelink==1.3.3
wheel==0.38.4
yarl==1.8.1
aiosqlite~=0.17.0
simpleeval~=0.9.12